Decouple GPGMM interface from impl.
