+++ src/gpgmm/common/Memory.h
@@ -17,55 +17,22 @@
 
 #include "gpgmm/common/Object.h"
 #include "gpgmm/utils/RefCount.h"
+#include "include/gpgmm.h"
 
 namespace gpgmm {
 
-    class MemoryPool;
-
-    /** \brief Represents a memory object.
-
-    When memory is sub-allocated, it will have a non-zero refcount.
-    */
     class MemoryBase : public ObjectBase {
       public:
-        /** \brief Constructs a memory object of the specified size and alignment.
-
-        @param size Size, in bytes, of the memory object.
-        @param alignment Alignment, in bytes, of the memory object.
-        */
         explicit MemoryBase(uint64_t size, uint64_t alignment);
         virtual ~MemoryBase() override;
 
-        /** \brief Return the size of the memory object.
-
-        \return Size, in bytes, of the memory object.
-        */
         uint64_t GetSize() const;
-
-        /** \brief Return the alignment of the memory object.
-
-        \return Alignment, in bytes, of the memory object.
-        */
         uint64_t GetAlignment() const;
 
-        /** \brief Get the memory pool managing the object.
+        IMemoryPool* GetPool() const;
+        void SetPool(IMemoryPool* pool);
 
-        \return A pointer to MemoryPool managing this memory object.
-        */
-        MemoryPool* GetPool() const;
-
-        /** \brief Set the memory pool to manage this object.
-
-        @param pool A pointer to MemoryPool used to manage this object.
-        */
-        void SetPool(MemoryPool* pool);
-
-        /** \brief Increments the sub-allocation reference count on the heap.
-         */
         void AddSubAllocationRef();
-
-        /** \brief Decrements the sub-allocation reference count on the heap.
-         */
         bool RemoveSubAllocationRef();
 
       private:
@@ -76,7 +43,7 @@ namespace gpgmm {
         const uint64_t mSize;
         const uint64_t mAlignment;
 
-        MemoryPool* mPool = nullptr;  // nullptr means no pool is assigned.
+        IMemoryPool* mPool = nullptr;  // nullptr means no pool is assigned.
     };
 
 }  // namespace gpgmm
