+++ src/gpgmm/d3d12/IUnknownImplD3D12.h
@@ -50,6 +50,14 @@ namespace gpgmm::d3d12 {
         GPGMM_REFCOUNT_TYPE mRefCount;  // Maintains the COM ref-count of this object.
     };
 
+// Helper to provide IUnknown-based object types the required overrides to use IUnknownImpl.
+#define DEFINE_IUNKNOWNIMPL_OVERRIDES()                                                       \
+    inline HRESULT STDMETHODCALLTYPE QueryInterface(REFIID riid, void** ppvObject) override { \
+        return IUnknownImpl::QueryInterface(riid, ppvObject);                                 \
+    }                                                                                         \
+    inline ULONG STDMETHODCALLTYPE AddRef() override { return IUnknownImpl::AddRef(); }       \
+    inline ULONG STDMETHODCALLTYPE Release() override { return IUnknownImpl::Release(); }
+
 }  // namespace gpgmm::d3d12
 
 #endif  // GPGMM_D3D12_IUNKNOWNIMPLD3D12_H_
