+++ src/gpgmm/d3d12/ResourceHeapAllocatorD3D12.cpp
@@ -26,14 +26,14 @@
 
 namespace gpgmm::d3d12 {
 
-    ResourceHeapAllocator::ResourceHeapAllocator(ResidencyManager* residencyManager,
+    ResourceHeapAllocator::ResourceHeapAllocator(IResidencyManager* residencyManager,
                                                  ID3D12Device* device,
                                                  D3D12_HEAP_PROPERTIES heapProperties,
                                                  D3D12_HEAP_FLAGS heapFlags)
         : mResidencyManager(residencyManager),
           mDevice(device),
           mHeapProperties(heapProperties),
-          mHeapFlags(heapFlags){
+          mHeapFlags(heapFlags) {
     }
 
     std::unique_ptr<MemoryAllocation> ResourceHeapAllocator::TryAllocateMemory(
@@ -55,14 +55,15 @@ namespace gpgmm::d3d12 {
         resourceHeapDesc.DebugName = L"Resource heap";
         resourceHeapDesc.Flags |= (mHeapFlags & D3D12_HEAP_FLAG_CREATE_NOT_RESIDENT) ? HEAPS_FLAG_NONE : HEAP_FLAG_ALWAYS_IN_BUDGET;
 
-        if (mResidencyManager != nullptr) {
+        ResidencyManager* residencyManager = static_cast<ResidencyManager*>(mResidencyManager);
+        if (residencyManager != nullptr) {
             resourceHeapDesc.MemorySegmentGroup = GetMemorySegmentGroup(
-                mHeapProperties.MemoryPoolPreference, mResidencyManager->IsUMA());
+                mHeapProperties.MemoryPoolPreference, residencyManager->IsUMA());
         }
 
-        Heap* resourceHeap = nullptr;
+        IHeap* resourceHeap = nullptr;
         if (FAILED(Heap::CreateHeap(
-                resourceHeapDesc, mResidencyManager,
+                resourceHeapDesc, residencyManager,
                 [&](ID3D12Pageable** ppPageableOut) -> HRESULT {
                     D3D12_HEAP_DESC heapDesc = {};
                     heapDesc.Properties = mHeapProperties;
