+++ src/mvi/gpgmm_d3d12_mvi.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "gpgmm_d3d12.h"
+#include "gpgmm_d3d12_mvi.h"
 
 #include <limits>
 
@@ -74,9 +74,9 @@ namespace gpgmm::d3d12 {
 
     // static
     HRESULT Heap::CreateHeap(const HEAP_DESC& descriptor,
-                             ResidencyManager* const pResidencyManager,
+                             IResidencyManager* const pResidencyManager,
                              CreateHeapFn&& createHeapFn,
-                             Heap** ppHeapOut) {
+                             IHeap** ppHeapOut) {
         Microsoft::WRL::ComPtr<ID3D12Pageable> pageable;
         ReturnIfFailed(createHeapFn(&pageable));
 
@@ -87,25 +87,78 @@ namespace gpgmm::d3d12 {
         return S_OK;
     }
 
-    HRESULT STDMETHODCALLTYPE Heap::QueryInterface(REFIID riid, void** ppvObject) {
-        return mPageable->QueryInterface(riid, ppvObject);
-    }
-
-    HEAP_INFO Heap::GetInfo() const {
-        return {};
-    }
-
     Heap::Heap(Microsoft::WRL::ComPtr<ID3D12Pageable> pageable,
                const HEAP_DESC& descriptor,
                bool isResidencyDisabled)
         : MemoryBase(descriptor.SizeInBytes, descriptor.Alignment), mPageable(std::move(pageable)) {
     }
 
+    HEAP_INFO Heap::GetInfo() const {
+        return {};
+    }
+
+    bool Heap::IsInResidencyLRUCacheForTesting() const {
+        return false;
+    }
+
+    bool Heap::IsResidencyLockedForTesting() const {
+        return false;
+    }
+
+    HRESULT STDMETHODCALLTYPE Heap::QueryInterface(REFIID riid, void** ppvObject) {
+        return mPageable->QueryInterface(riid, ppvObject);
+    }
+
+    ULONG STDMETHODCALLTYPE Heap::AddRef() {
+        return IUnknownImpl::AddRef();
+    }
+
+    ULONG STDMETHODCALLTYPE Heap::Release() {
+        return IUnknownImpl::Release();
+    }
+
+    uint64_t Heap::GetSize() const {
+        return MemoryBase::GetSize();
+    }
+
+    uint64_t Heap::GetAlignment() const {
+        return MemoryBase::GetAlignment();
+    }
+
+    void Heap::AddSubAllocationRef() {
+    }
+
+    bool Heap::RemoveSubAllocationRef() {
+        return true;
+    }
+
+    LPCWSTR Heap::GetDebugName() const {
+        return nullptr;
+    }
+
+    HRESULT Heap::SetDebugName(LPCWSTR Name) {
+        return E_NOTIMPL;
+    }
+
+    IMemoryPool* Heap::GetPool() const {
+        return nullptr;
+    }
+
+    void Heap::SetPool(IMemoryPool* pool) {
+    }
+
     // ResidencyList
+
+    HRESULT CreateResidencyList(IResidencyList** ppResidencyList) {
+        if (ppResidencyList != nullptr) {
+            *ppResidencyList = new ResidencyList();
+        }
+        return S_OK;
+    }
 
     ResidencyList::ResidencyList() = default;
 
-    HRESULT ResidencyList::Add(Heap* pHeap) {
+    HRESULT ResidencyList::Add(IHeap* pHeap) {
         return S_OK;
     }
 
@@ -113,11 +166,28 @@ namespace gpgmm::d3d12 {
         return S_OK;
     }
 
+    HRESULT STDMETHODCALLTYPE ResidencyList::QueryInterface(REFIID riid, void** ppvObject) {
+        return IUnknownImpl::QueryInterface(riid, ppvObject);
+    }
+
+    ULONG STDMETHODCALLTYPE ResidencyList::AddRef() {
+        return IUnknownImpl::AddRef();
+    }
+
+    ULONG STDMETHODCALLTYPE ResidencyList::Release() {
+        return IUnknownImpl::Release();
+    }
+
     // ResidencyManager
+
+    HRESULT CreateResidencyManager(const RESIDENCY_DESC& descriptor,
+                                   IResidencyManager** ppResidencyManagerOut) {
+        return ResidencyManager::CreateResidencyManager(descriptor, ppResidencyManagerOut);
+    }
 
     // static
     HRESULT ResidencyManager::CreateResidencyManager(const RESIDENCY_DESC& descriptor,
-                                                     ResidencyManager** ppResidencyManagerOut) {
+                                                     IResidencyManager** ppResidencyManagerOut) {
         if (ppResidencyManagerOut != nullptr) {
             *ppResidencyManagerOut = new ResidencyManager(descriptor);
         }
@@ -127,17 +197,17 @@ namespace gpgmm::d3d12 {
 
     ResidencyManager::~ResidencyManager() = default;
 
-    HRESULT ResidencyManager::LockHeap(Heap* pHeap) {
+    HRESULT ResidencyManager::LockHeap(IHeap* pHeap) {
         return S_OK;
     }
 
-    HRESULT ResidencyManager::UnlockHeap(Heap* pHeap) {
+    HRESULT ResidencyManager::UnlockHeap(IHeap* pHeap) {
         return S_OK;
     }
 
     HRESULT ResidencyManager::ExecuteCommandLists(ID3D12CommandQueue* pQueue,
                                                   ID3D12CommandList* const* ppCommandLists,
-                                                  ResidencyList* const* ppResidencyLists,
+                                                  IResidencyList* const* ppResidencyLists,
                                                   uint32_t count) {
         pQueue->ExecuteCommandLists(count, ppCommandLists);
         return S_OK;
@@ -162,6 +232,18 @@ namespace gpgmm::d3d12 {
 
     ResidencyManager::ResidencyManager(const RESIDENCY_DESC& descriptor)
         : mDevice(std::move(descriptor.Device)), mAdapter(std::move(descriptor.Adapter)) {
+    }
+
+    HRESULT STDMETHODCALLTYPE ResidencyManager::QueryInterface(REFIID riid, void** ppvObject) {
+        return IUnknownImpl::QueryInterface(riid, ppvObject);
+    }
+
+    ULONG STDMETHODCALLTYPE ResidencyManager::AddRef() {
+        return IUnknownImpl::AddRef();
+    }
+
+    ULONG STDMETHODCALLTYPE ResidencyManager::Release() {
+        return IUnknownImpl::Release();
     }
 
     // ResourceAllocation
@@ -196,29 +278,55 @@ namespace gpgmm::d3d12 {
         return {GetSize(), GetAlignment()};
     }
 
-    Heap* ResourceAllocation::GetMemory() const {
+    IHeap* ResourceAllocation::GetMemory() const {
         return static_cast<Heap*>(MemoryAllocation::GetMemory());
     }
 
     ResourceAllocation::ResourceAllocation(const RESOURCE_ALLOCATION_DESC& desc,
                                            MemoryAllocator* allocator,
-                                           Heap* resourceHeap,
+                                           IHeap* resourceHeap,
                                            Microsoft::WRL::ComPtr<ID3D12Resource> resource)
-        : MemoryAllocation(allocator, resourceHeap, desc.SizeInBytes),
-          mResource(std::move(resource)) {
+        : MemoryAllocation(allocator, resourceHeap), mResource(std::move(resource)) {
+    }
+
+    HRESULT STDMETHODCALLTYPE ResourceAllocation::QueryInterface(REFIID riid, void** ppvObject) {
+        return IUnknownImpl::QueryInterface(riid, ppvObject);
+    }
+
+    ULONG STDMETHODCALLTYPE ResourceAllocation::AddRef() {
+        return IUnknownImpl::AddRef();
+    }
+
+    ULONG STDMETHODCALLTYPE ResourceAllocation::Release() {
+        return IUnknownImpl::Release();
+    }
+
+    LPCWSTR ResourceAllocation::GetDebugName() const {
+        return nullptr;
+    }
+
+    HRESULT ResourceAllocation::SetDebugName(LPCWSTR Name) {
+        return E_NOTIMPL;
     }
 
     // ResourceAllocator
 
+    HRESULT CreateResourceAllocator(const ALLOCATOR_DESC& allocatorDescriptor,
+                                    IResourceAllocator** ppResourceAllocatorOut,
+                                    IResidencyManager** ppResidencyManagerOut) {
+        return ResourceAllocator::CreateResourceAllocator(
+            allocatorDescriptor, ppResourceAllocatorOut, ppResidencyManagerOut);
+    }
+
     // static
     HRESULT ResourceAllocator::CreateResourceAllocator(const ALLOCATOR_DESC& allocatorDescriptor,
-                                                       ResourceAllocator** ppResourceAllocatorOut,
-                                                       ResidencyManager** ppResidencyManagerOut) {
+                                                       IResourceAllocator** ppResourceAllocatorOut,
+                                                       IResidencyManager** ppResidencyManagerOut) {
         if (allocatorDescriptor.Device == nullptr || allocatorDescriptor.Adapter == nullptr) {
             return E_INVALIDARG;
         }
 
-        Microsoft::WRL::ComPtr<ResidencyManager> residencyManager;
+        Microsoft::WRL::ComPtr<IResidencyManager> residencyManager;
         if (ppResidencyManagerOut != nullptr) {
             RESIDENCY_DESC residencyDesc = {};
             residencyDesc.Device = allocatorDescriptor.Device;
@@ -234,7 +342,7 @@ namespace gpgmm::d3d12 {
                 ResidencyManager::CreateResidencyManager(residencyDesc, &residencyManager));
         }
 
-        Microsoft::WRL::ComPtr<ResourceAllocator> resourceAllocator;
+        Microsoft::WRL::ComPtr<IResourceAllocator> resourceAllocator;
         ReturnIfFailed(CreateResourceAllocator(allocatorDescriptor, residencyManager.Get(),
                                                &resourceAllocator));
 
@@ -250,9 +358,10 @@ namespace gpgmm::d3d12 {
     }
 
     // static
-    HRESULT ResourceAllocator::CreateResourceAllocator(const ALLOCATOR_DESC& allocatorDescriptor,
-                                                       ResidencyManager* pResidencyManager,
-                                                       ResourceAllocator** ppResourceAllocatorOut) {
+    HRESULT ResourceAllocator::CreateResourceAllocator(
+        const ALLOCATOR_DESC& allocatorDescriptor,
+        IResidencyManager* pResidencyManager,
+        IResourceAllocator** ppResourceAllocatorOut) {
         if (ppResourceAllocatorOut != nullptr) {
             *ppResourceAllocatorOut = new ResourceAllocator(allocatorDescriptor, pResidencyManager);
         }
@@ -264,8 +373,8 @@ namespace gpgmm::d3d12 {
                                               const D3D12_RESOURCE_DESC& resourceDescriptor,
                                               D3D12_RESOURCE_STATES initialResourceState,
                                               const D3D12_CLEAR_VALUE* pClearValue,
-                                              ResourceAllocation** ppResourceAllocationOut) {
-        Heap* resourceHeap = nullptr;
+                                              IResourceAllocation** ppResourceAllocationOut) {
+        IHeap* resourceHeap = nullptr;
         Microsoft::WRL::ComPtr<ID3D12Resource> committedResource;
 
         const D3D12_RESOURCE_ALLOCATION_INFO resourceInfo =
@@ -310,7 +419,7 @@ namespace gpgmm::d3d12 {
 
     HRESULT ResourceAllocator::CreateResource(
         Microsoft::WRL::ComPtr<ID3D12Resource> committedResource,
-        ResourceAllocation** ppResourceAllocationOut) {
+        IResourceAllocation** ppResourceAllocationOut) {
         return E_NOTIMPL;
     }
 
@@ -329,8 +438,8 @@ namespace gpgmm::d3d12 {
     }
 
     ResourceAllocator::ResourceAllocator(const ALLOCATOR_DESC& descriptor,
-                                         Microsoft::WRL::ComPtr<ResidencyManager> residencyManager)
-        : mDevice(std::move(descriptor.Device)), mResidencyManager(std::move(residencyManager)) {
+                                         IResidencyManager* pResidencyManager)
+        : mDevice(std::move(descriptor.Device)), mResidencyManager(pResidencyManager) {
     }
 
     void ResourceAllocator::DeallocateMemory(std::unique_ptr<MemoryAllocation> allocation) {
@@ -339,6 +448,18 @@ namespace gpgmm::d3d12 {
         mInfo.UsedMemoryCount--;
         mInfo.UsedBlockUsage -= allocationSize;
         delete allocation->GetMemory();
+    }
+
+    HRESULT STDMETHODCALLTYPE ResourceAllocator::QueryInterface(REFIID riid, void** ppvObject) {
+        return IUnknownImpl::QueryInterface(riid, ppvObject);
+    }
+
+    ULONG STDMETHODCALLTYPE ResourceAllocator::AddRef() {
+        return IUnknownImpl::AddRef();
+    }
+
+    ULONG STDMETHODCALLTYPE ResourceAllocator::Release() {
+        return IUnknownImpl::Release();
     }
 
 }  // namespace gpgmm::d3d12
