+++ src/tests/DummyMemoryAllocator.h
@@ -21,6 +21,37 @@
 
 namespace gpgmm {
 
+    class DummyMemory : public IMemoryObject, public MemoryBase {
+      public:
+        DummyMemory(uint64_t size, uint64_t alignment) : MemoryBase(size, alignment) {
+        }
+
+        // IMemoryObject
+        uint64_t GetSize() const override {
+            return MemoryBase::GetSize();
+        }
+
+        uint64_t GetAlignment() const override {
+            return MemoryBase::GetAlignment();
+        }
+
+        void AddSubAllocationRef() override {
+            return MemoryBase::AddSubAllocationRef();
+        }
+
+        bool RemoveSubAllocationRef() override {
+            return MemoryBase::RemoveSubAllocationRef();
+        }
+
+        IMemoryPool* GetPool() const override {
+            return MemoryBase::GetPool();
+        }
+
+        void SetPool(IMemoryPool* pool) override {
+            return MemoryBase::SetPool(pool);
+        }
+    };
+
     class DummyMemoryAllocator : public MemoryAllocator {
       public:
         DummyMemoryAllocator() = default;
@@ -43,7 +74,7 @@ namespace gpgmm {
             mInfo.UsedMemoryUsage += request.SizeInBytes;
 
             return std::make_unique<MemoryAllocation>(
-                this, new MemoryBase(request.SizeInBytes, request.Alignment), request.SizeInBytes);
+                this, new DummyMemory(request.SizeInBytes, request.Alignment), request.SizeInBytes);
         }
 
         void DeallocateMemory(std::unique_ptr<MemoryAllocation> allocation) override {
