+++ src/tests/end2end/D3D12ResidencyManagerTests.cpp
@@ -68,7 +68,7 @@ class D3D12ResidencyManagerTests : public D3D12TestBase, public ::testing::Test 
         return residencyDesc;
     }
 
-    uint64_t GetBudgetLeft(ResidencyManager* residencyManager,
+    uint64_t GetBudgetLeft(IResidencyManager* residencyManager,
                            const DXGI_MEMORY_SEGMENT_GROUP& memorySegmentGroup) {
         DXGI_QUERY_VIDEO_MEMORY_INFO segment = {};
         residencyManager->QueryVideoMemoryInfo(memorySegmentGroup, &segment);
@@ -81,9 +81,9 @@ TEST_F(D3D12ResidencyManagerTests, CreateResourceHeapNotResident) {
     // Adapters that do not support creating heaps will ignore D3D12_HEAP_FLAG_CREATE_NOT_RESIDENT.
     GPGMM_SKIP_TEST_IF(!mCaps->IsCreateHeapNotResidentSupported());
 
-    ComPtr<ResidencyManager> residencyManager;
-    ASSERT_SUCCEEDED(ResidencyManager::CreateResidencyManager(
-        CreateBasicResidencyDesc(kDefaultBudget), &residencyManager));
+    ComPtr<IResidencyManager> residencyManager;
+    ASSERT_SUCCEEDED(
+        CreateResidencyManager(CreateBasicResidencyDesc(kDefaultBudget), &residencyManager));
 
     constexpr uint64_t kHeapSize = GPGMM_MB_TO_BYTES(10);
 
@@ -113,14 +113,14 @@ TEST_F(D3D12ResidencyManagerTests, CreateResourceHeapNotResident) {
         return S_OK;
     };
 
-    ASSERT_FAILED(Heap::CreateHeap(resourceHeapAlwaysInBudgetDesc, residencyManager.Get(),
-                                   createHeapNotResidentFn, nullptr));
+    ASSERT_FAILED(CreateHeap(resourceHeapAlwaysInBudgetDesc, residencyManager.Get(),
+                             createHeapNotResidentFn, nullptr));
 }
 
 TEST_F(D3D12ResidencyManagerTests, CreateResourceHeap) {
-    ComPtr<ResidencyManager> residencyManager;
-    ASSERT_SUCCEEDED(ResidencyManager::CreateResidencyManager(
-        CreateBasicResidencyDesc(kDefaultBudget), &residencyManager));
+    ComPtr<IResidencyManager> residencyManager;
+    ASSERT_SUCCEEDED(
+        CreateResidencyManager(CreateBasicResidencyDesc(kDefaultBudget), &residencyManager));
 
     constexpr uint64_t kHeapSize = GPGMM_MB_TO_BYTES(10);
 
@@ -152,15 +152,13 @@ TEST_F(D3D12ResidencyManagerTests, CreateResourceHeap) {
     resourceHeapDesc.SizeInBytes = kHeapSize;
     resourceHeapDesc.MemorySegmentGroup = DXGI_MEMORY_SEGMENT_GROUP_LOCAL;
 
-    ASSERT_FAILED(
-        Heap::CreateHeap(resourceHeapDesc, residencyManager.Get(), badCreateHeapFn, nullptr));
+    ASSERT_FAILED(CreateHeap(resourceHeapDesc, residencyManager.Get(), badCreateHeapFn, nullptr));
 
-    ASSERT_SUCCEEDED(
-        Heap::CreateHeap(resourceHeapDesc, residencyManager.Get(), createHeapFn, nullptr));
+    ASSERT_SUCCEEDED(CreateHeap(resourceHeapDesc, residencyManager.Get(), createHeapFn, nullptr));
 
     // Create a resource heap without residency.
-    ComPtr<Heap> resourceHeap;
-    ASSERT_SUCCEEDED(Heap::CreateHeap(resourceHeapDesc, nullptr, createHeapFn, &resourceHeap));
+    ComPtr<IHeap> resourceHeap;
+    ASSERT_SUCCEEDED(CreateHeap(resourceHeapDesc, nullptr, createHeapFn, &resourceHeap));
 
     // Ensure the unmanaged resource heap state is always unknown. Even though D3D12 implicitly
     // creates heaps as resident, untrack resource heaps would never transition out from
@@ -170,7 +168,7 @@ TEST_F(D3D12ResidencyManagerTests, CreateResourceHeap) {
 
     // Create a resource heap with residency.
     ASSERT_SUCCEEDED(
-        Heap::CreateHeap(resourceHeapDesc, residencyManager.Get(), createHeapFn, &resourceHeap));
+        CreateHeap(resourceHeapDesc, residencyManager.Get(), createHeapFn, &resourceHeap));
     ASSERT_NE(resourceHeap, nullptr);
 
     EXPECT_EQ(resourceHeap->GetInfo().Status, gpgmm::d3d12::RESIDENCY_STATUS_CURRENT_RESIDENT);
@@ -205,9 +203,9 @@ TEST_F(D3D12ResidencyManagerTests, CreateResourceHeap) {
 }
 
 TEST_F(D3D12ResidencyManagerTests, CreateDescriptorHeap) {
-    ComPtr<ResidencyManager> residencyManager;
-    ASSERT_SUCCEEDED(ResidencyManager::CreateResidencyManager(
-        CreateBasicResidencyDesc(kDefaultBudget), &residencyManager));
+    ComPtr<IResidencyManager> residencyManager;
+    ASSERT_SUCCEEDED(
+        CreateResidencyManager(CreateBasicResidencyDesc(kDefaultBudget), &residencyManager));
 
     D3D12_DESCRIPTOR_HEAP_DESC heapDesc = {};
     heapDesc.Type = D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV;
@@ -228,9 +226,9 @@ TEST_F(D3D12ResidencyManagerTests, CreateDescriptorHeap) {
         return S_OK;
     };
 
-    ComPtr<Heap> descriptorHeap;
-    ASSERT_SUCCEEDED(Heap::CreateHeap(descriptorHeapDesc, residencyManager.Get(), createHeapFn,
-                                      &descriptorHeap));
+    ComPtr<IHeap> descriptorHeap;
+    ASSERT_SUCCEEDED(
+        CreateHeap(descriptorHeapDesc, residencyManager.Get(), createHeapFn, &descriptorHeap));
 
     EXPECT_EQ(descriptorHeap->GetInfo().Status, gpgmm::d3d12::RESIDENCY_STATUS_UNKNOWN);
     EXPECT_EQ(descriptorHeap->GetInfo().IsLocked, false);
@@ -265,38 +263,46 @@ TEST_F(D3D12ResidencyManagerTests, CreateDescriptorHeap) {
 }
 
 TEST_F(D3D12ResidencyManagerTests, CreateResidencyList) {
-    ComPtr<ResourceAllocator> resourceAllocator;
-    ASSERT_SUCCEEDED(ResourceAllocator::CreateResourceAllocator(CreateBasicAllocatorDesc(),
-                                                                &resourceAllocator, nullptr));
+    ComPtr<IResourceAllocator> resourceAllocator;
+    ASSERT_SUCCEEDED(
+        CreateResourceAllocator(CreateBasicAllocatorDesc(), &resourceAllocator, nullptr));
 
     ALLOCATION_DESC allocationDesc = {};
     allocationDesc.HeapType = D3D12_HEAP_TYPE_DEFAULT;
 
-    ComPtr<ResourceAllocation> allocation;
+    ComPtr<IResourceAllocation> allocation;
     ASSERT_SUCCEEDED(resourceAllocator->CreateResource(allocationDesc, CreateBasicBufferDesc(1),
                                                        D3D12_RESOURCE_STATE_COMMON, nullptr,
                                                        &allocation));
 
     // Inserting a non-existant heap should always fail
     {
-        ResidencyList list;
-        Heap* invalid = nullptr;
-        ASSERT_FAILED(list.Add(invalid));
+        ComPtr<IResidencyList> list;
+        ASSERT_SUCCEEDED(CreateResidencyList(&list));
+        IHeap* invalid = nullptr;
+        ASSERT_FAILED(list->Add(invalid));
     }
 
     // Inserting from a valid allocation should always succeed.
     {
-        ResidencyList list;
-        ASSERT_SUCCEEDED(list.Add(allocation->GetMemory()));
-        ASSERT_SUCCEEDED(list.Add(allocation->GetMemory()));
+        ComPtr<IResidencyList> list;
+        ASSERT_SUCCEEDED(CreateResidencyList(&list));
+
+        ASSERT_SUCCEEDED(list->Add(allocation->GetMemory()));
+        ASSERT_SUCCEEDED(list->Add(allocation->GetMemory()));
     }
 
     // Re-inserting allocation between two sets should always succeed.
     {
-        ResidencyList listA;
-        ASSERT_SUCCEEDED(listA.Add(allocation->GetMemory()));
-        ResidencyList listB(listA);
-        EXPECT_EQ(listB.Add(allocation->GetMemory()), S_OK);
+        ComPtr<IResidencyList> listA;
+        ASSERT_SUCCEEDED(CreateResidencyList(&listA));
+
+        ASSERT_SUCCEEDED(listA->Add(allocation->GetMemory()));
+
+        ComPtr<IResidencyList> listB;
+        ASSERT_SUCCEEDED(CreateResidencyList(&listB));
+
+        EXPECT_EQ(listB->Add(allocation->GetMemory()), S_OK);
     }
 }
 
@@ -306,7 +312,7 @@ TEST_F(D3D12ResidencyManagerTests, CreateResidencyManager) {
         RESIDENCY_DESC residencyDesc = CreateBasicResidencyDesc(kDefaultBudget);
         residencyDesc.Adapter = nullptr;
 
-        ASSERT_FAILED(ResidencyManager::CreateResidencyManager(residencyDesc, nullptr));
+        ASSERT_FAILED(CreateResidencyManager(residencyDesc, nullptr));
     }
 
     // Create residency without device must always fail.
@@ -314,35 +320,34 @@ TEST_F(D3D12ResidencyManagerTests, CreateResidencyManager) {
         RESIDENCY_DESC residencyDesc = CreateBasicResidencyDesc(kDefaultBudget);
         residencyDesc.Device = nullptr;
 
-        ASSERT_FAILED(ResidencyManager::CreateResidencyManager(residencyDesc, nullptr));
+        ASSERT_FAILED(CreateResidencyManager(residencyDesc, nullptr));
     }
 
     // Create residency alone.
     {
-        ComPtr<ResidencyManager> residencyManager;
-        ASSERT_SUCCEEDED(ResidencyManager::CreateResidencyManager(
-            CreateBasicResidencyDesc(kDefaultBudget), nullptr));
+        ComPtr<IResidencyManager> residencyManager;
+        ASSERT_SUCCEEDED(CreateResidencyManager(CreateBasicResidencyDesc(kDefaultBudget), nullptr));
     }
 
     // Create allocator with residency support, together.
     {
-        ComPtr<ResidencyManager> residencyManager;
-        ComPtr<ResourceAllocator> resourceAllocator;
-        ASSERT_SUCCEEDED(ResourceAllocator::CreateResourceAllocator(
-            CreateBasicAllocatorDesc(), &resourceAllocator, &residencyManager));
+        ComPtr<IResidencyManager> residencyManager;
+        ComPtr<IResourceAllocator> resourceAllocator;
+        ASSERT_SUCCEEDED(CreateResourceAllocator(CreateBasicAllocatorDesc(), &resourceAllocator,
+                                                 &residencyManager));
         EXPECT_NE(resourceAllocator, nullptr);
         EXPECT_NE(residencyManager, nullptr);
     }
 
     // Create allocator with residency, seperately.
     {
-        ComPtr<ResidencyManager> residencyManager;
-        ASSERT_SUCCEEDED(ResidencyManager::CreateResidencyManager(
-            CreateBasicResidencyDesc(kDefaultBudget), &residencyManager));
+        ComPtr<IResidencyManager> residencyManager;
+        ASSERT_SUCCEEDED(
+            CreateResidencyManager(CreateBasicResidencyDesc(kDefaultBudget), &residencyManager));
 
-        ComPtr<ResourceAllocator> resourceAllocator;
-        ASSERT_SUCCEEDED(ResourceAllocator::CreateResourceAllocator(
-            CreateBasicAllocatorDesc(), residencyManager.Get(), &resourceAllocator));
+        ComPtr<IResourceAllocator> resourceAllocator;
+        ASSERT_SUCCEEDED(CreateResourceAllocator(CreateBasicAllocatorDesc(), residencyManager.Get(),
+                                                 &resourceAllocator));
         EXPECT_NE(resourceAllocator, nullptr);
         EXPECT_NE(residencyManager, nullptr);
     }
@@ -353,21 +358,19 @@ TEST_F(D3D12ResidencyManagerTests, CreateResidencyManagerNoLeak) {
 
     // Create allocator with residency support, together.
     {
-        ComPtr<ResidencyManager> residencyManager;
-        ComPtr<ResourceAllocator> resourceAllocator;
-        ResourceAllocator::CreateResourceAllocator(CreateBasicAllocatorDesc(), &resourceAllocator,
-                                                   &residencyManager);
+        ComPtr<IResidencyManager> residencyManager;
+        ComPtr<IResourceAllocator> resourceAllocator;
+        CreateResourceAllocator(CreateBasicAllocatorDesc(), &resourceAllocator, &residencyManager);
     }
 
     // Create allocator with residency, seperately.
     {
-        ComPtr<ResidencyManager> residencyManager;
-        ResidencyManager::CreateResidencyManager(CreateBasicResidencyDesc(kDefaultBudget),
-                                                 &residencyManager);
+        ComPtr<IResidencyManager> residencyManager;
+        CreateResidencyManager(CreateBasicResidencyDesc(kDefaultBudget), &residencyManager);
 
-        ComPtr<ResourceAllocator> resourceAllocator;
-        ResourceAllocator::CreateResourceAllocator(CreateBasicAllocatorDesc(),
-                                                   residencyManager.Get(), &resourceAllocator);
+        ComPtr<IResourceAllocator> resourceAllocator;
+        CreateResourceAllocator(CreateBasicAllocatorDesc(), residencyManager.Get(),
+                                &resourceAllocator);
     }
 
     GPGMM_TEST_MEMORY_LEAK_END();
@@ -378,12 +381,12 @@ TEST_F(D3D12ResidencyManagerTests, CreateResidencyManagerNoLeak) {
 TEST_F(D3D12ResidencyManagerTests, OverBudget) {
     RESIDENCY_DESC residencyDesc = CreateBasicResidencyDesc(kDefaultBudget);
 
-    ComPtr<ResidencyManager> residencyManager;
-    ASSERT_SUCCEEDED(ResidencyManager::CreateResidencyManager(residencyDesc, &residencyManager));
+    ComPtr<IResidencyManager> residencyManager;
+    ASSERT_SUCCEEDED(CreateResidencyManager(residencyDesc, &residencyManager));
 
-    ComPtr<ResourceAllocator> resourceAllocator;
-    ASSERT_SUCCEEDED(ResourceAllocator::CreateResourceAllocator(
-        CreateBasicAllocatorDesc(), residencyManager.Get(), &resourceAllocator));
+    ComPtr<IResourceAllocator> resourceAllocator;
+    ASSERT_SUCCEEDED(CreateResourceAllocator(CreateBasicAllocatorDesc(), residencyManager.Get(),
+                                             &resourceAllocator));
 
     constexpr uint64_t kBufferMemorySize = GPGMM_MB_TO_BYTES(1);
     const D3D12_RESOURCE_DESC bufferDesc = CreateBasicBufferDesc(kBufferMemorySize);
@@ -392,9 +395,9 @@ TEST_F(D3D12ResidencyManagerTests, OverBudget) {
     bufferAllocationDesc.HeapType = D3D12_HEAP_TYPE_DEFAULT;
 
     // Keep allocating until we reach the budget.
-    std::vector<ComPtr<ResourceAllocation>> allocationsBelowBudget = {};
+    std::vector<ComPtr<IResourceAllocation>> allocationsBelowBudget = {};
     while (resourceAllocator->GetInfo().UsedMemoryUsage + kBufferMemorySize <= kDefaultBudget) {
-        ComPtr<ResourceAllocation> allocation;
+        ComPtr<IResourceAllocation> allocation;
         ASSERT_SUCCEEDED(resourceAllocator->CreateResource(
             bufferAllocationDesc, bufferDesc, D3D12_RESOURCE_STATE_COMMON, nullptr, &allocation));
         allocationsBelowBudget.push_back(std::move(allocation));
@@ -409,11 +412,11 @@ TEST_F(D3D12ResidencyManagerTests, OverBudget) {
     constexpr uint64_t kMemoryOverBudget = GPGMM_MB_TO_BYTES(10);
 
     // Allocating the same amount over budget, where older allocations will be evicted.
-    std::vector<ComPtr<ResourceAllocation>> allocationsAboveBudget = {};
+    std::vector<ComPtr<IResourceAllocation>> allocationsAboveBudget = {};
     const uint64_t currentMemoryUsage = resourceAllocator->GetInfo().UsedMemoryUsage;
 
     while (currentMemoryUsage + kMemoryOverBudget > resourceAllocator->GetInfo().UsedMemoryUsage) {
-        ComPtr<ResourceAllocation> allocation;
+        ComPtr<IResourceAllocation> allocation;
         ASSERT_SUCCEEDED(resourceAllocator->CreateResource(
             bufferAllocationDesc, bufferDesc, D3D12_RESOURCE_STATE_COMMON, nullptr, &allocation));
         allocationsAboveBudget.push_back(std::move(allocation));
@@ -436,12 +439,12 @@ TEST_F(D3D12ResidencyManagerTests, OverBudgetAsync) {
     RESIDENCY_DESC residencyDesc = CreateBasicResidencyDesc(kBudgetIsDeterminedByOS);
     residencyDesc.Flags ^= RESIDENCY_FLAG_NEVER_UPDATE_BUDGET_ON_WORKER_THREAD;
 
-    ComPtr<ResidencyManager> residencyManager;
-    ASSERT_SUCCEEDED(ResidencyManager::CreateResidencyManager(residencyDesc, &residencyManager));
+    ComPtr<IResidencyManager> residencyManager;
+    ASSERT_SUCCEEDED(CreateResidencyManager(residencyDesc, &residencyManager));
 
-    ComPtr<ResourceAllocator> resourceAllocator;
-    ASSERT_SUCCEEDED(ResourceAllocator::CreateResourceAllocator(
-        CreateBasicAllocatorDesc(), residencyManager.Get(), &resourceAllocator));
+    ComPtr<IResourceAllocator> resourceAllocator;
+    ASSERT_SUCCEEDED(CreateResourceAllocator(CreateBasicAllocatorDesc(), residencyManager.Get(),
+                                             &resourceAllocator));
 
     constexpr uint64_t kBufferMemorySize = GPGMM_MB_TO_BYTES(1);
     const D3D12_RESOURCE_DESC bufferDesc = CreateBasicBufferDesc(kBufferMemorySize);
@@ -459,10 +462,10 @@ TEST_F(D3D12ResidencyManagerTests, OverBudgetAsync) {
 
     // Keep allocating until we reach the budget. Should a budget change occur, we must also
     // terminate the loop since we cannot guarantee all allocations will be created resident.
-    std::vector<ComPtr<ResourceAllocation>> allocations = {};
+    std::vector<ComPtr<IResourceAllocation>> allocations = {};
     while (resourceAllocator->GetInfo().UsedMemoryUsage + kBufferMemorySize < memoryUnderBudget &&
            GetBudgetLeft(residencyManager.Get(), bufferMemorySegment) >= kBufferMemorySize) {
-        ComPtr<ResourceAllocation> allocation;
+        ComPtr<IResourceAllocation> allocation;
         ASSERT_SUCCEEDED(resourceAllocator->CreateResource(
             bufferAllocationDesc, bufferDesc, D3D12_RESOURCE_STATE_COMMON, nullptr, &allocation));
 
@@ -480,18 +483,18 @@ TEST_F(D3D12ResidencyManagerTests, OverBudgetAsync) {
 TEST_F(D3D12ResidencyManagerTests, OverBudgetDisablesGrowth) {
     RESIDENCY_DESC residencyDesc = CreateBasicResidencyDesc(kDefaultBudget);
 
-    ComPtr<ResidencyManager> residencyManager;
-    ASSERT_SUCCEEDED(ResidencyManager::CreateResidencyManager(residencyDesc, &residencyManager));
+    ComPtr<IResidencyManager> residencyManager;
+    ASSERT_SUCCEEDED(CreateResidencyManager(residencyDesc, &residencyManager));
 
     ALLOCATOR_DESC allocatorDesc = CreateBasicAllocatorDesc();
     allocatorDesc.MemoryGrowthFactor = 2;
 
-    ComPtr<ResourceAllocator> resourceAllocator;
-    ASSERT_SUCCEEDED(ResourceAllocator::CreateResourceAllocator(
-        CreateBasicAllocatorDesc(), residencyManager.Get(), &resourceAllocator));
+    ComPtr<IResourceAllocator> resourceAllocator;
+    ASSERT_SUCCEEDED(CreateResourceAllocator(CreateBasicAllocatorDesc(), residencyManager.Get(),
+                                             &resourceAllocator));
 
-    std::vector<ComPtr<ResourceAllocation>> allocations = {};
-    std::vector<Heap*> resourceHeaps = {};
+    std::vector<ComPtr<IResourceAllocation>> allocations = {};
+    std::vector<IHeap*> resourceHeaps = {};
 
     constexpr uint64_t kBufferMemorySize = GPGMM_MB_TO_BYTES(1);
     const D3D12_RESOURCE_DESC bufferDesc = CreateBasicBufferDesc(kBufferMemorySize);
@@ -500,7 +503,7 @@ TEST_F(D3D12ResidencyManagerTests, OverBudgetDisablesGrowth) {
     bufferAllocationDesc.HeapType = D3D12_HEAP_TYPE_DEFAULT;
 
     while (resourceAllocator->GetInfo().UsedMemoryUsage + kBufferMemorySize <= kDefaultBudget) {
-        ComPtr<ResourceAllocation> allocation;
+        ComPtr<IResourceAllocation> allocation;
         ASSERT_SUCCEEDED(resourceAllocator->CreateResource(
             bufferAllocationDesc, bufferDesc, D3D12_RESOURCE_STATE_COMMON, nullptr, &allocation));
 
@@ -524,12 +527,12 @@ TEST_F(D3D12ResidencyManagerTests, OverBudgetDisablesGrowth) {
 TEST_F(D3D12ResidencyManagerTests, OverBudgetWithLockedHeaps) {
     RESIDENCY_DESC residencyDesc = CreateBasicResidencyDesc(kDefaultBudget);
 
-    ComPtr<ResidencyManager> residencyManager;
-    ASSERT_SUCCEEDED(ResidencyManager::CreateResidencyManager(residencyDesc, &residencyManager));
+    ComPtr<IResidencyManager> residencyManager;
+    ASSERT_SUCCEEDED(CreateResidencyManager(residencyDesc, &residencyManager));
 
-    ComPtr<ResourceAllocator> resourceAllocator;
-    ASSERT_SUCCEEDED(ResourceAllocator::CreateResourceAllocator(
-        CreateBasicAllocatorDesc(), residencyManager.Get(), &resourceAllocator));
+    ComPtr<IResourceAllocator> resourceAllocator;
+    ASSERT_SUCCEEDED(CreateResourceAllocator(CreateBasicAllocatorDesc(), residencyManager.Get(),
+                                             &resourceAllocator));
 
     constexpr uint64_t kBufferMemorySize = GPGMM_MB_TO_BYTES(1);
     const D3D12_RESOURCE_DESC bufferDesc = CreateBasicBufferDesc(kBufferMemorySize);
@@ -538,9 +541,9 @@ TEST_F(D3D12ResidencyManagerTests, OverBudgetWithLockedHeaps) {
     bufferAllocationDesc.HeapType = D3D12_HEAP_TYPE_DEFAULT;
 
     // Keep allocating until we reach the budget.
-    std::vector<ComPtr<ResourceAllocation>> allocationsBelowBudget = {};
+    std::vector<ComPtr<IResourceAllocation>> allocationsBelowBudget = {};
     while (resourceAllocator->GetInfo().UsedMemoryUsage + kBufferMemorySize <= kDefaultBudget) {
-        ComPtr<ResourceAllocation> allocation;
+        ComPtr<IResourceAllocation> allocation;
         ASSERT_SUCCEEDED(resourceAllocator->CreateResource(
             bufferAllocationDesc, bufferDesc, D3D12_RESOURCE_STATE_COMMON, nullptr, &allocation));
 
@@ -572,21 +575,21 @@ TEST_F(D3D12ResidencyManagerTests, OverBudgetWithLockedHeaps) {
 // swaps the residency status using ExecuteCommandList: first set gets paged-in again, second set
 // gets paged-out.
 TEST_F(D3D12ResidencyManagerTests, ExecuteCommandListOverBudget) {
-    ComPtr<ResidencyManager> residencyManager;
-    ASSERT_SUCCEEDED(ResidencyManager::CreateResidencyManager(
-        CreateBasicResidencyDesc(kDefaultBudget), &residencyManager));
+    ComPtr<IResidencyManager> residencyManager;
+    ASSERT_SUCCEEDED(
+        CreateResidencyManager(CreateBasicResidencyDesc(kDefaultBudget), &residencyManager));
 
-    ComPtr<ResourceAllocator> resourceAllocator;
-    ASSERT_SUCCEEDED(ResourceAllocator::CreateResourceAllocator(
-        CreateBasicAllocatorDesc(), residencyManager.Get(), &resourceAllocator));
+    ComPtr<IResourceAllocator> resourceAllocator;
+    ASSERT_SUCCEEDED(CreateResourceAllocator(CreateBasicAllocatorDesc(), residencyManager.Get(),
+                                             &resourceAllocator));
 
     constexpr uint64_t kBufferMemorySize = GPGMM_MB_TO_BYTES(1);
     const D3D12_RESOURCE_DESC bufferDesc = CreateBasicBufferDesc(kBufferMemorySize);
 
     // Create the first set of heaps below the budget.
-    std::vector<ComPtr<ResourceAllocation>> firstSetOfHeaps = {};
+    std::vector<ComPtr<IResourceAllocation>> firstSetOfHeaps = {};
     while (resourceAllocator->GetInfo().UsedMemoryUsage + kBufferMemorySize <= kDefaultBudget) {
-        ComPtr<ResourceAllocation> allocation;
+        ComPtr<IResourceAllocation> allocation;
         ASSERT_SUCCEEDED(resourceAllocator->CreateResource(
             {}, bufferDesc, D3D12_RESOURCE_STATE_COMMON, nullptr, &allocation));
         EXPECT_EQ(allocation->GetMemory()->GetInfo().Status, RESIDENCY_STATUS_CURRENT_RESIDENT);
@@ -594,9 +597,9 @@ TEST_F(D3D12ResidencyManagerTests, ExecuteCommandListOverBudget) {
     }
 
     // Create the second set of heaps above the budget, the first set will be evicted.
-    std::vector<ComPtr<ResourceAllocation>> secondSetOfHeaps = {};
+    std::vector<ComPtr<IResourceAllocation>> secondSetOfHeaps = {};
     for (uint64_t i = 0; i < kDefaultBudget / kBufferMemorySize; i++) {
-        ComPtr<ResourceAllocation> allocation;
+        ComPtr<IResourceAllocation> allocation;
         ASSERT_SUCCEEDED(resourceAllocator->CreateResource(
             {}, bufferDesc, D3D12_RESOURCE_STATE_COMMON, nullptr, &allocation));
         EXPECT_EQ(allocation->GetMemory()->GetInfo().Status, RESIDENCY_STATUS_CURRENT_RESIDENT);
@@ -604,9 +607,11 @@ TEST_F(D3D12ResidencyManagerTests, ExecuteCommandListOverBudget) {
     }
 
     // Page-in the first set of heaps using ExecuteCommandLists (and page-out the second set).
-    ResidencyList firstSetOfHeapsWorkingSet;
+    ComPtr<IResidencyList> firstSetOfHeapsWorkingSet;
+    ASSERT_SUCCEEDED(CreateResidencyList(&firstSetOfHeapsWorkingSet));
+
     for (auto& allocation : firstSetOfHeaps) {
-        firstSetOfHeapsWorkingSet.Add(allocation->GetMemory());
+        firstSetOfHeapsWorkingSet->Add(allocation->GetMemory());
     }
 
     ComPtr<ID3D12CommandAllocator> commandAllocator;
@@ -623,7 +628,7 @@ TEST_F(D3D12ResidencyManagerTests, ExecuteCommandListOverBudget) {
     ASSERT_SUCCEEDED(mDevice->CreateCommandQueue(&queueDesc, IID_PPV_ARGS(&queue)));
 
     {
-        ResidencyList* residencyLists[] = {&firstSetOfHeapsWorkingSet};
+        IResidencyList* residencyLists[] = {firstSetOfHeapsWorkingSet.Get()};
         ID3D12CommandList* commandLists[] = {commandList.Get()};
         ASSERT_SUCCEEDED(
             residencyManager->ExecuteCommandLists(queue.Get(), commandLists, residencyLists, 1));
@@ -640,13 +645,15 @@ TEST_F(D3D12ResidencyManagerTests, ExecuteCommandListOverBudget) {
     }
 
     // Page-in the second set of heaps using ExecuteCommandLists (and page-out the first set).
-    ResidencyList secondSetOfHeapsWorkingSet;
+    ComPtr<IResidencyList> secondSetOfHeapsWorkingSet;
+    ASSERT_SUCCEEDED(CreateResidencyList(&secondSetOfHeapsWorkingSet));
+
     for (auto& allocation : secondSetOfHeaps) {
-        secondSetOfHeapsWorkingSet.Add(allocation->GetMemory());
+        secondSetOfHeapsWorkingSet->Add(allocation->GetMemory());
     }
 
     {
-        ResidencyList* residencyLists[] = {&secondSetOfHeapsWorkingSet};
+        IResidencyList* residencyLists[] = {secondSetOfHeapsWorkingSet.Get()};
         ID3D12CommandList* commandLists[] = {commandList.Get()};
         ASSERT_SUCCEEDED(
             residencyManager->ExecuteCommandLists(queue.Get(), commandLists, residencyLists, 1));
